+++ packages/world/src/item/stack.ts
@@ -9,9 +9,9 @@ import {
 } from "@serenityjs/nbt";
 
 import { type ItemComponent, ItemTagComponent } from "../components";
-import { BlockContainer, EntityContainer, type Container } from "../container";
 
-import type { World } from "../world";
+import type { Container } from "../container";
+import type { Dimension, World } from "../world";
 import type {
 	NetworkItemInstanceDescriptor,
 	NetworkItemStackDescriptor
@@ -54,6 +54,11 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	public readonly maxAmount: number;
 
 	/**
+	 * The dimension the item is in.
+	 */
+	public dimension: Dimension;
+
+	/**
 	 * The container of the item stack.
 	 */
 	public container: Container | null = null;
@@ -68,16 +73,60 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	 * @param identifier The identifier of the item.
 	 * @param amount The amount of the item.
 	 * @param metadata The metadata of the item.
+	 * @param dimension The dimension the item is in.
 	 */
-	public constructor(identifier: T, amount: number, metadata?: number) {
+	public constructor(
+		identifier: T,
+		amount: number,
+		metadata: number,
+		dimension: Dimension
+	) {
 		this.type = ItemType.get(identifier) as ItemType<T>;
-		this.metadata = metadata ?? 0;
+		this.amount = amount;
+		this.metadata = metadata;
+		this.dimension = dimension;
 		this.stackable = this.type.stackable;
 		this.maxAmount = this.type.maxAmount;
-		this.amount = amount;
 
-		// Update the item stack.
-		this.update();
+		// Get the components of the entity from the entity palette
+		const components = this.getWorld().items.getRegistry(identifier);
+
+		// Register the type components to the entity.
+		for (const component of components) {
+			// Skip if the entity already has the component
+			if (this.components.has(component.identifier)) continue;
+
+			// Create a new instance of the component
+			new component(this, component.identifier);
+		}
+
+		// Register the tag components to the item.
+		for (const tag of this.type.tags) {
+			// Get the component from the registry
+			const component = this.getWorld()
+				.items.getAllComponents()
+				.find((x) => {
+					// If the identifier is undefined, we will skip it.
+					if (!x.identifier || !(x.prototype instanceof ItemTagComponent))
+						return false;
+
+					// Initialize the component as a BlockStateComponent.
+					const component = x as typeof ItemTagComponent;
+
+					// Check if the identifier includes the key.
+					// As some states dont include a namespace.
+					return component.tag === tag;
+				});
+
+			// Check if the component exists.
+			if (!component) continue;
+
+			// Check if the item already has the component.
+			if (this.components.has(component.identifier)) continue;
+
+			// Create the component.
+			new component(this, component.identifier);
+		}
 	}
 
 	/**
@@ -90,6 +139,14 @@ class ItemStack<T extends keyof Items = keyof Items> {
 
 		// Update the item in the container.
 		this.update();
+	}
+
+	/**
+	 * Get the world the item is in.
+	 * @returns The world the item is in.
+	 */
+	public getWorld(): World {
+		return this.dimension.world;
 	}
 
 	/**
@@ -110,50 +167,6 @@ class ItemStack<T extends keyof Items = keyof Items> {
 
 		// Set the item in the container.
 		else this.container.setItem(slot, this);
-	}
-
-	// TODO: Get rid of this...
-	public sync(): void {
-		// Check if the item is in a world.
-		if (!this.getWorld()) return;
-
-		// Get the world the item is in.
-		const world = this.getWorld() as World;
-
-		// Register the type components to the item.
-		for (const component of world.items.getRegistry(this.type.identifier)) {
-			// Skip if the item already has the component.
-			if (this.components.has(component.identifier)) continue;
-
-			// Create the component.
-			new component(this, component.identifier);
-		}
-
-		// Register the tag components to the item.
-		for (const tag of this.type.tags) {
-			// Get the component from the registry
-			const component = world.items.getAllComponents().find((x) => {
-				// If the identifier is undefined, we will skip it.
-				if (!x.identifier || !(x.prototype instanceof ItemTagComponent))
-					return false;
-
-				// Initialize the component as a BlockStateComponent.
-				const component = x as typeof ItemTagComponent;
-
-				// Check if the identifier includes the key.
-				// As some states dont include a namespace.
-				return component.tag === tag;
-			});
-
-			// Check if the component exists.
-			if (!component) continue;
-
-			// Check if the item already has the component.
-			if (this.components.has(component.identifier)) continue;
-
-			// Create the component.
-			new component(this, component.identifier);
-		}
 	}
 
 	/**
@@ -284,26 +297,6 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	}
 
 	/**
-	 * Gets the world the item is in.
-	 * @returns The world the item is in.
-	 */
-	public getWorld(): World | null {
-		// Check if the item is in a container.
-		if (!this.container) return null;
-
-		// Check if the container is an entity container.
-		if (this.container instanceof EntityContainer)
-			return this.container.entity.dimension.world;
-
-		// Check if the container is a block container.
-		if (this.container instanceof BlockContainer)
-			return this.container.block.dimension.world;
-
-		// Return null if the world is not found.
-		return null;
-	}
-
-	/**
 	 * Converts the item stack to a network item instance descriptor.
 	 * Which is used on the protocol level.
 	 * @param item The item stack to convert.
@@ -332,10 +325,12 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	/**
 	 * Converts a network item instance descriptor to an item stack.
 	 * @param descriptor The network item instance descriptor.
+	 * @param dimension The dimension the item is in.
 	 * @returns The item stack.
 	 */
 	public static fromNetworkInstance(
-		descriptor: NetworkItemInstanceDescriptor
+		descriptor: NetworkItemInstanceDescriptor,
+		dimension: Dimension
 	): ItemStack | null {
 		// Get the item type from the network.
 		const type = ItemType.getByNetwork(descriptor.network);
@@ -347,7 +342,8 @@ class ItemStack<T extends keyof Items = keyof Items> {
 		const item = new ItemStack(
 			type.identifier,
 			descriptor.stackSize ?? 1,
-			descriptor.metadata ?? 1
+			descriptor.metadata ?? 1,
+			dimension
 		);
 
 		// Return the item stack.
@@ -376,14 +372,16 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	 * @param type The type of the item.
 	 * @param amount The amount of the item.
 	 * @param metadata The metadata of the item.
+	 * @param dimension The dimension the item is in.
 	 * @returns The item stack.
 	 */
 	public static create<T extends keyof Items>(
 		type: ItemType<T>,
 		amount: number,
-		metadata?: number
+		metadata: number,
+		dimension: Dimension
 	): ItemStack<T> {
-		return new ItemStack(type.identifier, amount, metadata);
+		return new ItemStack(type.identifier, amount, metadata, dimension);
 	}
 
 	/**
@@ -395,10 +393,6 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	 * @returns The serialized representation of the item stack.
 	 */
 	public static serialize(itemStack: ItemStack, tagName?: string): CompoundTag {
-		const world = itemStack.getWorld();
-
-		if (!world) throw new Error(`ItemStack is not in a world`);
-
 		// Create the root tag.
 		const root = new CompoundTag(tagName).addTag(
 			new StringTag("Name", itemStack.type.identifier),
@@ -419,7 +413,9 @@ class ItemStack<T extends keyof Items = keyof Items> {
 		// Iterate over the components and serialize them.
 		for (const component of itemStack.getComponents()) {
 			// Get the component type.
-			const type = world.items.getComponent(component.identifier);
+			const type = itemStack
+				.getWorld()
+				.items.getComponent(component.identifier);
 			if (!type) continue;
 
 			// Create a data compound tag for the data to be written to.
@@ -448,10 +444,11 @@ class ItemStack<T extends keyof Items = keyof Items> {
 	 * and constructs an ItemStack object from it.
 	 *
 	 * @param tag The CompoundTag containing the serialized item stack data.
+	 * @param dimension The dimension the item is in.
 	 * @returns The deserialized Itemstack
 	 * @throws Will throw an error if the tag does not contain a valid item identifier.
 	 */
-	public static deserialize(world: World, tag: CompoundTag): ItemStack {
+	public static deserialize(dimension: Dimension, tag: CompoundTag): ItemStack {
 		// Create the item stack.
 		const itemIdentifier = tag.getTag<StringTag>("Name");
 
@@ -459,10 +456,11 @@ class ItemStack<T extends keyof Items = keyof Items> {
 		if (!itemIdentifier) throw new Error(`Invalid item tag`);
 
 		// Creat the item stack.
-		const item = new ItemStack(
+		const itemStack = new ItemStack(
 			itemIdentifier.value as keyof Items,
 			tag.getTag<ByteTag>("Count")?.value ?? 1,
-			tag.getTag<ShortTag>("Damage")?.value ?? 0
+			tag.getTag<ShortTag>("Damage")?.value ?? 0,
+			dimension
 		);
 
 		// Deserialize the components.
@@ -477,15 +475,15 @@ class ItemStack<T extends keyof Items = keyof Items> {
 			if (!data) continue;
 
 			// Get the component type.
-			const type = world.items.getComponent(identifier);
+			const type = dimension.world.items.getComponent(identifier);
 			if (!type) continue;
 
 			// Deserialize the component.
-			type.deserialize(data, item);
+			type.deserialize(data, itemStack);
 		}
 
 		// Return the item stack.
-		return item;
+		return itemStack;
 	}
 }