+++ packages/world/src/player/player.ts
@@ -299,6 +299,9 @@ class Player extends Entity {
 		// Trigger the onSpawn method of all applicable components
 		for (const component of this.getComponents()) component.onSpawn?.();
 
+		// Set the player's alive state
+		this.isAlive = true;
+
 		// Sync the players data
 		return this.sync();
 	}
@@ -488,6 +491,7 @@ class Player extends Entity {
 				packet.dimension = dimension.type;
 				packet.position = position;
 				packet.respawn = true;
+				packet.hasLoadingScreen = false;
 
 				// Send the packet to the player
 				this.session.sendImmediate(packet);