+++ packages/jest-presets/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @serenityjs/jest-presets
 
+## 0.5.1
+
+### Patch Changes
+
+- 2880acc: init v0.5.1-beta
+
 ## 0.5.0
 
 ### Minor Changes