+++ packages/network/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @serenityjs/network
 
+## 0.5.1
+
+### Patch Changes
+
+- 2880acc: init v0.5.1-beta
+- Updated dependencies [2880acc]
+  - @serenityjs/emitter@0.5.1
+  - @serenityjs/logger@0.5.1
+  - @serenityjs/protocol@0.5.1
+  - @serenityjs/raknet@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes